+++ src/c_cmds.c
@@ -760,7 +760,7 @@ consolecmd_t consolecmds[] =
     CCMD(play, "", "", play_cmd_func1, play_cmd_func2, true, PLAYCMDFORMAT,
         "Plays a " BOLDITALICS("sound effect") " or " BOLDITALICS("music") " lump."),
     CVAR_INT(playergender, "", "", playergender_cvar_func1, playergender_cvar_func2, CF_NONE, PLAYERGENDERVALUEALIAS,
-        "The player's gender (" BOLD("male") ", " BOLD("female") " or " BOLD("other") ")."),
+        "The player's gender (" BOLD("male") ", " BOLD("female") " or " BOLD("nonbinary") ")."),
     CVAR_STR(playername, "", "", null_func1, playername_cvar_func2, CF_NONE,
         "The player's name."),
     CCMD(playerstats, "", "", null_func1, playerstats_cmd_func2, false, "",